Remove route constraint
Don't use constraints for input validation. If constraints are used for input validation, invalid input results in a 404 Not Found response. Route constraints are used to disambiguate similar routes, not to validate the inputs for a particular route. Invalid input should produce a 400 Bad Request with an appropriate error message.